+++ model/src/main/java/com/century/logregator/dao/MvnTagDao.java
@@ -1,14 +1,8 @@
 package com.century.logregator.dao;
 
-import com.century.logregator.model.Application;
 import com.century.logregator.model.MvnTag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.Collection;
 
 public class MvnTagDao {
     private static final String INSERT_MVN_TAG_SQL = "INSERT INTO mvn_tag(group_id, artifact_id, version)" +
@@ -25,7 +19,7 @@ public class MvnTagDao {
         if(mvnTag.getId() != null){
             throw new IllegalArgumentException("Cant save mvnTag twice, update is not possible");
         }
-        Long id = jdbcTemplate.queryForObject(INSERT_MVN_TAG_SQL, Long.class, mvnTag.getGroupId(), mvnTag.getArtifactId(), mvnTag.getVersion());
+        Integer id = jdbcTemplate.queryForObject(INSERT_MVN_TAG_SQL, Integer.class, mvnTag.getGroupId(), mvnTag.getArtifactId(), mvnTag.getVersion());
         mvnTag.setId(id);
     }
