+++ model/src/main/java/com/century/logregator/model/JarInfo.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @EqualsAndHashCode
 public class JarInfo {
     @Getter @Setter
-    private Long id;
+    private int id;
 
     @Getter @Setter
     private String fileName;
@@ -24,7 +24,7 @@ public class JarInfo {
     @Getter @Setter
     private MvnTag mvnTag;
 
-    public Long getMvnTagId(){
+    public Integer getMvnTagId(){
         if(mvnTag == null){
             return null;
         }
@@ -34,7 +34,7 @@ public class JarInfo {
      * Application, this jar file belongs to
      */
     @Getter
-    Long applicationId;
+    Integer applicationId;
 
     public JarInfo(String fileName, MvnTag mvnTag) {
         this.fileName = fileName;
